8d250f9 FIX: warnings from gcc in Linux

+++ Source/igstkBinaryData.h
@@ -57,7 +57,7 @@ public:
   virtual ~BinaryData();
 
   /** SetSize method resizes/allocates memory */
-  void SetSize(unsigned int size);
+  void SetSize(size_t size);
 
   /** GetSize method returns the size of data */
   size_t GetSize() const;
@@ -69,7 +69,7 @@ public:
   ContainerType& GetData() { return this->m_Data; }
 
   /** Copy data from an array */
-  void CopyFrom(unsigned char* inputBegin, unsigned int inputLength);
+  void CopyFrom(unsigned char* inputBegin, size_t inputLength);
 
   /** Copy data into an array */
   void CopyTo(unsigned char* output) const;
@@ -78,7 +78,7 @@ public:
   void Append(unsigned char byte);
   
   /** Append data from an array */
-  void Append(const unsigned char* inputBegin, unsigned int inputLength);
+  void Append(const unsigned char* inputBegin, size_t inputLength);
 
   /** Assign the values of one BinaryData to another */
   const BinaryData & operator=( const BinaryData & inputBinaryData );
